refactor: combine GK Text components into one
